apiProject/requirements.txt
allure-python-commons==2.13.5
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
distlib==0.3.9
et_xmlfile==2.0.0
filelock==3.18.0
fonttools==4.56.0
h11==0.14.0
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
numpy==2.2.4
pluggy==1.5.0
